requirements.txt
absl-py==0.7.1
astor==0.7.1
attrs==19.1.0
backcall==0.1.0
bleach==1.5.0
cffi==1.12.2
chardet==3.0.4
Click==7.0
cloudpickle==0.8.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.9
entrypoints==0.3
-e git+https://github.com/meta-inf/experiments@a278b3cb11c2e3bd3ec08b0bc18712b2f4a21749#egg=experiments
future==0.17.1
gast==0.2.0
glfw==1.7.1
google-pasta==0.1.7
grpcio==1.19.0
gym==0.12.1
h5py==2.9.0
html5lib==0.9999999
idna==2.8
imageio==2.4.1
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
jedi==0.13.3
Jinja2==2.10
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.0.1
lockfile==0.12.2
Markdown==3.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
mock==2.0.0
mpi4py==3.0.0
msgpack==0.6.1
msgpack-numpy==0.4.4.3
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.2
notebook==5.7.8
numpy==1.16.4
opencv-python==4.0.0.21
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.4.1
progressbar2==3.39.3
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.7.1
psutil==5.6.1
ptyprocess==0.6.0
pycparser==2.19
pyglet==1.3.2
Pygments==2.3.1
PyOpenGL==3.1.0
pyparsing==2.3.1
pyrsistent==0.14.11
python-dateutil==2.8.0
python-utils==2.3.0
pytz==2018.9
PyWavelets==1.0.2
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.3
requests==2.21.0
scikit-image==0.13.0
scikit-learn==0.20.3
scipy==1.3.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.11.0
tabulate==0.8.3
tensorboard==1.14.0
#tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorpack==0.9.5
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
torch==0.3.1
torchvision==0.2.0
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
urllib3==1.24.1
wcwidth==0.1.7
Werkzeug==0.14.1
widgetsnbextension==3.4.2
wrapt==1.11.1
-e git+https://github.com/thu-ml/zhusuan@4932e225331c7ff615ec54cad0d115fa1e371955#egg=zhusuan
zmq==0.0.0
